product edit fix

+++ app/Http/Controllers/Products/ProductController.php
@@ -332,25 +332,27 @@ class ProductController extends Controller
                     if($variation['id'] != 0) {
                         $variation_model = Product::find($variation['id']);
                         if(!$variation_model) $not_saved_products_id[] = $variation['id'];
-                        $variation_model->update([
-                            // 'info_id' => $product->id,
-                            'price' => intval($variation['price']),
-                            'is_popular' => $variation['is_popular'],
-                            'product_of_the_day' => $variation['product_of_the_day'],
-                            'status' => $variation['status'],
-                            'slug' => $this->product_slug_create($product, $additional_for_slug, $variation_model->id)
-                        ]); // model, c_id, is_available ne izpolzuyetsya
+                        if($variation_model) {
+                            $variation_model->update([
+                                // 'info_id' => $product->id,
+                                'price' => intval($variation['price']),
+                                'is_popular' => $variation['is_popular'],
+                                'product_of_the_day' => $variation['product_of_the_day'],
+                                'status' => $variation['status'],
+                                'slug' => $this->product_slug_create($product, $additional_for_slug, $variation_model->id)
+                            ]); // model, c_id, is_available ne izpolzuyetsya
 
-                        /*
-                         * sync images
-                         */
-                        $qolgan_rasmlar_ids = array_map(function($i) {
-                            return $i['id'];
-                        }, $qolgan_rasmlar);
-                        /*
-                         * kerakmas rasmlarni o'chiramiz
-                         */
-                        $variation_model->images()->whereNotIn('product_images.id', $qolgan_rasmlar_ids)->delete();
+                            /*
+                             * sync images
+                             */
+                            $qolgan_rasmlar_ids = array_map(function($i) {
+                                return $i['id'];
+                            }, $qolgan_rasmlar);
+                            /*
+                             * kerakmas rasmlarni o'chiramiz
+                             */
+                            $variation_model->images()->whereNotIn('product_images.id', $qolgan_rasmlar_ids)->delete();
+                        }
                     } else {
                         $variation_model = Product::create([
                             'info_id' => $product->id,
